requirements.txt
Flask==3.0.3
supabase==2.4.1
python-dotenv==1.0.1
bcrypt==4.1.2
